# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/group/exit": {
      "get": {
        "description": "room_id token 为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "退出群聊接口",
        "parameters": [
          {
            "type": "string",
            "description": "群号",
            "name": "room_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"msg\":\"退出成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/group/group": {
      "post": {
        "description": "info token 为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "创建群聊接口",
        "parameters": [
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "群名称",
            "name": "info",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200,\"msg\": \"创建群聊成功,群号为:8660920\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/group/grouplist": {
      "get": {
        "description": "room_id  为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "获取群成员列表接口",
        "parameters": [
          {
            "type": "string",
            "description": "群号",
            "name": "room_id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200,\"data\": {\t\"data\": [\t{\t\"Indently\": \"6a2a462c-a107-48ea-82e5-74e308327e6f\",\t\"Username\": \"admin\", \"Password\": \"21232f297a57a5a743894a0e4a801fc3\",\t\"Use_status\": 0,\t\"Register_time\": \"2023-03-13 17:05:08\",\t\"Email\": \"3096407764@qq.com\",\t\"account\": \"3169387148\"\t},\t{\t\t\"Indently\": \"cacda2d3-4a77-4afa-94b5-6ff2c036d126\", \"Username\": \"leilong\",\t\t   \"Password\": \"e10adc3949ba59abbe56e057f20f883e\", \"Use_status\": 0, \"Register_time\": \"2023-03-12 19:07:16\",\"Email\": \"3096407768@qq.com\", \"account\": \"9546270155\"}]\t}, \"msg\": \"获取数据成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/group/join": {
      "get": {
        "description": "room_id token 为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "加入群聊接口",
        "parameters": [
          {
            "type": "string",
            "description": "群号",
            "name": "room_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"msg\": \"加入群聊成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/delete": {
      "get": {
        "description": "token 账号为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "删除好友接口",
        "parameters": [
          {
            "type": "string",
            "description": "账号",
            "name": "account",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"msg\":\"删除成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "description": "上传文件\nfile 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "上传文件接口",
        "parameters": [
          {
            "type": "string",
            "description": "表单name",
            "name": "file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":1,\"msg\":\"\\u0001个文件上传成功\",\"url\":\"127.0.0.1:8080/img/12.png\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/friend_list": {
      "get": {
        "description": "token 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "好友列表接口",
        "parameters": [
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{ \"code\": 200, \"data\": {\"data\": [    {\"Indently\": \"6a2a462c-a107-48ea-82e5-74e308327e6f\", \"Username\": \"admin\", \"Password\": \"21232f297a57a5a743894a0e4a801fc3\", \"Use_status\": 0, \"Register_time\": \"2023-03-13 17:05:08\",\"Email\": \"3096407764@qq.com\", \"account\": \"3169387148\"}]}, \"msg\": \"获取数据成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/get_message": {
      "get": {
        "description": "room_id token 为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "获取聊天记录接口",
        "parameters": [
          {
            "type": "string",
            "description": "房间号",
            "name": "room_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":1,\"data\":{\"data\":[{\"id\":29,\"idently\":\"6a2a462c-a107-48ea-82e5-74e308327e6f\",\"message_id\":6858759,\"message\":\"你好，我是张三\",\"room_idently\":\"0820018\",\"messagesend_time\":1679052325},{\"id\":30,\"idently\":\"cacda2d3-4a77-4afa-94b5-6ff2c036d126\",\"message_id\":72843315,\"message\":\"你好，我是李四\",\"room_idently\":\"0820018\",\"messagesend_time\":1679052347},{\"id\":31,\"idently\":\"cacda2d3-4a77-4afa-94b5-6ff2c036d126\",\"message_id\":91900639,\"message\":\"你好，我是李四\",\"room_idently\":\"0820018\",\"messagesend_time\":1679054157},{\"id\":32,\"idently\":\"cacda2d3-4a77-4afa-94b5-6ff2c036d126\",\"message_id\":63367923,\"message\":\"你好，我是李四\",\"room_idently\":\"0820018\",\"messagesend_time\":1679054157},{\"id\":33,\"idently\":\"cacda2d3-4a77-4afa-94b5-6ff2c036d126\",\"message_id\":30613339,\"message\":\"你好，我是李四\",\"room_idently\":\"0820018\",\"messagesend_time\":1679054158}]},\"msg\":\"获取数据成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/join": {
      "get": {
        "description": "账号 token 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "添加好友接口",
        "parameters": [
          {
            "type": "string",
            "description": "账号",
            "name": "account",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"msg\":\"添加好友成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "用户名 密码 验证码为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "登录接口",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "验证码",
            "name": "code",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"msg\": \"登陆成功\", \"token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbmRlbnRseSI6IjZhMmE0NjJjLWExMDctNDhlYS04MmU1LTc0ZTMwODMyN2U2ZiIsInVzZXJuYW1lIjoiYWRtaW4iLCJpc3MiOiJ0ZXN0IiwiZXhwIjoxNjc4Nzg2NTM1fQ.P4dJ_f2UGhKbpiIqHxTxghRKwKIlCpF2XjryHCSnKKk\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/online": {
      "get": {
        "description": "account 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "获取好友在线状态接口",
        "parameters": [
          {
            "type": "string",
            "description": "account",
            "name": "token",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{ \"code\": 200, \"msg\": \"获取用户状态成功！\", \"status\": true }",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "description": "用户名 密码 邮箱为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "注册接口",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "邮箱",
            "name": "email",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"account\": \"3169387148\", \"code\": 200, \"msg\": \"注册成功\"}",
            "schema": {
              "type": ""
            }
          }
        }
      }
    },
    "/user/send_code": {
      "get": {
        "description": "用户名 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "验证码接口",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{ \"code\": \"530757\", \"msg\": \"获取验证码成功！\" }",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/userinfo": {
      "get": {
        "description": "token 为必填",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "获取用户信息接口",
        "parameters": [
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\":200,\"data\":{\"data\":{\"Indently\":\"949c770f-0b1e-4ca0-a15a-927ee5396c3b\",\"Username\":\"aadmin\",\"Password\":\"21232f297a57a5a743894a0e4a801fc3\",\"Use_status\":0,\"Register_time\":\"2023-03-12 19:07:30\",\"Email\":\"3096407769@qq.com\",\"account\":\"8356511203\"}},\"msg\":\"获取数据成功！\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/websocket": {
      "get": {
        "description": "token 为必填",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "公共方法"
        ],
        "summary": "websocket连接接口",
        "parameters": [
          {
            "type": "string",
            "description": "token",
            "name": "token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {}
      }
    }
  }
}